Fixed edge cases for delete category

+++ controllers/categoryController.js
@@ -17,7 +17,16 @@ exports.createCategory = async (req, res) => {
 exports.getAllCategories = async (req, res) => {
   try {
     const categories = await Category.find();
-    res.status(200).json(categories);
+    const categoriesWithBlogCount = await Promise.all(
+      categories.map(async (category) => {
+        const blogCount = await Blog.countDocuments({ category: category._id });
+        return {
+          ...category.toObject(),
+          blogCount,
+        };
+      })
+    );
+    res.status(200).json(categoriesWithBlogCount);
   } catch (err) {
     res.status(500).json({ error: "Failed to retrieve categories." });
   }
@@ -50,6 +59,13 @@ exports.updateCategory = async (req, res) => {
 exports.deleteCategoryByName = async (req, res) => {
   try {
     const { categoryName } = req.params;
+    const categoryToDelete = await Category.findOne({ categoryName: categoryName.toLowerCase() });
+    if (!categoryToDelete) return res.status(404).json({ error: "Category not found." });
+    let unassignedCategory = await Category.findOne({ categoryName: "Unassigned" });
+    await Blog.updateMany(
+      { category: categoryToDelete._id },
+      { $set: { category: unassignedCategory._id } }
+    );
     const category = await Category.findOneAndDelete({ categoryName: categoryName.toLowerCase() });
     if (!category) return res.status(404).json({ error: "Category not found." });
     res.status(200).json({ message: "Category deleted successfully." });
